// src/modules/output/interfaces.ts
export interface OutputHelperOptionsConfigTableStyles {
    [name: string]: OutputHelperOptionsConfigTableStyle

    FAT?: OutputHelperOptionsConfigTableStyle
    SLIM?: OutputHelperOptionsConfigTableStyle
    NONE?: OutputHelperOptionsConfigTableStyle
}

export interface OutputHelperOptionsConfigTableStyle {
    [name: string]: string,

    'top'         ?: string,
    'top-mid'     ?: string,
    'top-left'    ?: string,
    'top-right'   ?: string,
    'bottom'      ?: string,
    'bottom-mid'  ?: string,
    'bottom-left' ?: string,
    'bottom-right'?: string,
    'left'        ?: string,
    'left-mid'    ?: string,
    'mid'         ?: string,
    'mid-mid'     ?: string,
    'right'       ?: string,
    'right-mid'   ?: string,
    'middle'      ?: string,
}

export interface OutputHelperOptionsConfig extends HelperOptionsConfig {
    quiet?: boolean,
    colors?: boolean,
    options?: {
        quiet?: HelperOptionsConfigOption,
        colors?: HelperOptionsConfigOption
    },
    resetOnNewline?: boolean,
    styles?: { [name: string]: string },
    tableStyle?: OutputHelperOptionsConfigTableStyles
}